Created a method 'likeEscapeCaluse' for the EOSQLExpression as previously the escape clause was appended in EOKeyValueQualifier.sqlStringForSQLExpression: using a literal.  The escape clause and whether or not it is needed / used is really adaptor dependent. So we need the EOSQLExpression method so that the behaviour may be more easily modified at the adaptor level.
I then went ahead and overrode that behaviour for the MySQL adaptor as the escape character itself needs to be escaped when we define it.  I'll admit this is very flaky as if MySQL is running in NO_BACKSLASH_ESCAPES mode then it should NOT be escaped.  I can't think of any way to deal with that.